moved encoder back into ftcsv
